microservices changed, fixes

+++ ms-loan/src/main/java/com/prestabanco/loan/repository/LoanRepository.java
@@ -4,6 +4,8 @@ import com.prestabanco.loan.entity.Loan;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface LoanRepository extends JpaRepository<Loan, Integer> {
+public interface LoanRepository extends JpaRepository<Loan, Long> {
 }
